Activate more tests
